move txn verify to data/transactions/verify/
