+++ node/node.go
@@ -36,6 +36,7 @@ import (
 	"github.com/algorand/go-algorand/data/bookkeeping"
 	"github.com/algorand/go-algorand/data/pools"
 	"github.com/algorand/go-algorand/data/transactions"
+	"github.com/algorand/go-algorand/data/transactions/verify"
 	"github.com/algorand/go-algorand/ledger"
 	"github.com/algorand/go-algorand/logging"
 	"github.com/algorand/go-algorand/network"
@@ -401,7 +402,7 @@ func (node *AlgorandFullNode) BroadcastSignedTxGroup(txgroup []transactions.Sign
 	proto := config.Consensus[b.CurrentProtocol]
 
 	for _, tx := range txgroup {
-		err = ledger.TxnVerify(&tx, spec, proto)
+		err = verify.Txn(&tx, spec, proto)
 		if err != nil {
 			node.log.Warnf("malformed transaction: %v - transaction was %+v", err, tx)
 			return err
